Remove type guards in Python for received messages
Received messages are constructed with Haskell, which is by definition
type safe after successful compilation, so there is no need to double
check something that GHC has already ensured. Changes in the
Serialization API will indeed force a change in the Python code, but it
has to be changed regardless if type guards are used. So all the type
guards does is to make the code harder to read and increase line noise.
If I must use Haskell, then might as well maximize the proportion of
code that is written in Haskell under the guarantee of type safety,
which means minimizing the number of Python code.

Mypy is still used to type check responses, because there is no compiler
to guarantee its type safety here.